refactor(backend): 修改 updatePostStatistics 参数类型
- 将 PostService 接口中的 updatePostStatistics 方法的参数类型从 Long 改为 Integer
- 该修改统一了代码风格，因为前端传参多为 Integer 类型

+++ backend/src/main/java/cn/hmck/service/PostService.java
@@ -53,5 +53,5 @@ public interface PostService extends IService<Post> {
     Post getPostById(Integer id);
 
     // 更新动态的点赞数和评论数
-    void updatePostStatistics(Long postId);
+    void updatePostStatistics(Integer postId);
 }
